Combine BaseTest and EndpointBaseTest as only EndpointBaseTest is used
